+++ app/src/main/java/com/cajetan/youtubeplus/VideoListAdapter.kt
@@ -133,7 +133,7 @@ class VideoListAdapter(videos: List<Video>, listener: ListItemClickListener, con
 
             val duration = video.contentDetails.duration
             videoDurationView.text = when (duration) {
-                "PT0S" -> {
+                mContext.getString(R.string.live_video_duration) -> {
                     videoDurationView.setBackgroundColor(Color.RED)
                     videoDurationView.setTextColor(Color.WHITE)
                     mContext.getString(R.string.live)
@@ -150,7 +150,7 @@ class VideoListAdapter(videos: List<Video>, listener: ListItemClickListener, con
                 else -> ""
             }
 
-            if (thumbnailUrl != "") {
+            if (thumbnailUrl.isNotEmpty()) {
                 Picasso.get().load(thumbnailUrl)
                         .resize(dpToPixel(160f, mContext), dpToPixel(90f, mContext))
                         .centerCrop().into(videoThumbnailView)