+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 15.01.2019
 
+* [PlayerActivity] Refactoring and improvements 
 * [PlayerActivity] Fixed switching to a new video on a new share Intent
 
 #### 14.01.2019